[feat/#31] 회원가입 페이지 설정 변경(이메일+성별 통합)
