+++ app/(auth)/components/StepThemes.tsx
@@ -33,7 +33,7 @@ export default function StepThemes({ onNext, onBack }: Props) {
 
   return (
     <div>
-      <h2 className="text-h2 font-bold mb-6">선호하는 게임 테마를 선택해주세요</h2>
+      <h2 className="text-body text-font-100 font-semibold mb-2">선호하는 게임 테마를 선택해주세요</h2>
 
       <div className="grid grid-cols-2 sm:grid-cols-3 gap-4 mb-4">
         {THEMES.map((theme) => (
